fix: always abort the maintenance service
I hit this bug when one the API calls got hanging, and submitting the
machine config with `apply-config` never takes the node out of
maintenance mode, as `.GracefulStop()` may hang forever waiting for all
the calls to finish.

This way we always abort at some timeout and stop the server forcefully.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>