fix: kill the task processes when cleaning up stale task
The bug was triggered by `containerd` crash (restart), in this case
runner receives an error as if the process exited.
Runner tries to restart the container, but as the container is still
running, attempt to delete the task would fail.

With this change Talos always tries to kill the running container and
waits for the container to terminate.

The error message when the bug was triggered looks like:

```
service[kubelet](Waiting): Error running Containerd(kubelet), going to restart forever: failed to clean up task "kubelet": task must be stopped before deletion: running: failed precondition
```

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>