fix: don't report link name in route statuses
This is a band-aid specifically for the `release-1.3` branch, there will be
a proper fix in the `main` branch developed.

The problem shows up with Calico CNI, as every pod starting/stopping
results in multiple updates to link-local resources (for the routes to
IPv6 localhost address) which are repeatedly attached to different
`caliXXXX` links. The rate of updates is so high that COSI runtime can't
correctly keep up with the event stream leading to runtime crash.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>