fix: disable Wireless Lan using dtoverlay
Talos has no wireless support & wireless kernel drivers,
so disabling it the recommended way might actually might save power consumption.

It could save ~45 mA:
https://forums.raspberrypi.com/viewtopic.php?t=257144#p1568474

Or 'The WiFi half of the wireless chip will be powered but be held in reset':
https://forums.raspberrypi.com/viewtopic.php?t=343854#p2060246

Either way, it does not hurt and it should be treated the same as bluetooth.

Signed-off-by: Nico Berlee <nico.berlee@on2it.net>
Signed-off-by: Noel Georgi <git@frezbo.dev>