fix: allow `talosctl cp` to handle special files in `/proc`
There is some refactoring to simplify things, but mostly handle files
which report size 0 in `stat`, but actually contain data when read.

We try to read up to the small buffer, if we read whole file, we use
that as contents, otherwise we still skip the file, as we need to write
tar header with size _before_ we read the whole file.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>