fix: properly skip/cleanup controlplane configs for workers
This bug is pretty cosmetic, but it shows up as a wrong check when
performing worker upgrade - Talos pretends it checks e.g. kube-apiserver
version which doesn't make sense for workers.

There were two bugs in the code:

* check for machine type was done against `TypeWorker`, while
  `MachineType` resource is initially created as `TypeUnknown`
* the cleanup code was not implemented

As I touched the code, I updated controller and tests to use modern
conventions.

Signed-off-by: Andrey Smirnov <andrey.smirnov@talos-systems.com>