fix: restore compatibility with Kubernetes 1.26
The `--container-runtime-endpoint` flag was moved to the kubelet config
only in Kubernetes 1.27.

Kubernetes 1.27 and 1.28 still support the flag (even though 1.28 shows
the warning that the flag is deprecated).

This is direct commit to `release-1.5`, as `main` is going to use
Kubernetes 1.27-1.29, so it should be using a config file.

Fixes #7648

Signed-off-by: Andrey Smirnov <andrey.smirnov@siderolabs.com>