docs: fix broken links in metal-network-configuration.md
Fixed the set of same links in 1.4, 1.5, 1.6, and 1.7, with an exception of a link in 1.4 where the it links to boot assets and boot assets, if we were to place a copy in that version, is missing a bunch of supporting links. Opted to skip that update, as that documentation is unsupported.

Signed-off-by: edwinavalos <edwin.a.avalos@gmail.com>