fix: cancel all operations if `c.V1Alpha2().Run` fails
The current code contains a deadlock if `c.V1Alpha2().Run` fails with an error because `EnforceKSPPRequirements`
will indefinitely wait for a condition to happen, but the controller runtime is not running.
This commit addresses this by replacing the error channel with `context.WithCancelCause` and
canceling the context if the controller runner fails.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>