fix: merge PodSecurity admission configs without duplicate exemptions
When we strategically merge admission plugin configs of type "PodSecurity", we were concatenating the fields `namespaces`, `runtimeClasses` and `usernames`, which are string slices.

This did not pass the validation in the kube-apiserver and put it into a crash loop with errors like:
```
run.go:74] "command failed" err="failed to apply admission: failed to initialize admission plugin \"PodSecurity\": PodSecurity invalid: exemptions.namespaces[1]: Duplicate value: \"kube-system\"
```

Fix it by treating `PodSecurity` specially and not duplicating the elements in these fields while merging.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>